Simpler cookie mechanism Added a WCP definition
