+++ website/.vscode/settings.json
@@ -6,7 +6,7 @@
 		"**/CVS": true,
 		"**/.DS_Store": true,
 		"**/Thumbs.db": true,
-		"node_modules": true,
+		"node_modules": false,
 		"dist": true
 	},
 	"editor.formatOnSave": true,
