+++ website/public/PlayFabAdminApi.js
@@ -14,9 +14,9 @@ if(!PlayFab.settings) {
 if(!PlayFab._internalSettings) {
     PlayFab._internalSettings = {
         entityToken: null,
-        sdkVersion: "1.180.240913",
+        sdkVersion: "1.182.241025",
         requestGetParams: {
-            sdk: "JavaScriptSDK-1.180.240913"
+            sdk: "JavaScriptSDK-1.182.241025"
         },
         sessionTicket: null,
         verticalName: null, // The name of a customer vertical. This is only for customers running a private cluster. Generally you shouldn't touch this
@@ -224,7 +224,7 @@ if(!PlayFab._internalSettings) {
 }
 
 PlayFab.buildIdentifier = "adobuild_javascriptsdk_114";
-PlayFab.sdkVersion = "1.180.240913";
+PlayFab.sdkVersion = "1.182.241025";
 PlayFab.GenerateErrorReport = function (error) {
     if (error == null)
         return "";
