+++ website/tsconfig.json
@@ -22,6 +22,6 @@
 	},
 	"include": ["src"],
 	"exclude": ["node_modules"],
-	"files": ["./src/index.d.ts"],
+	"files": ["./src/index.d.ts", "./src/wcp.d.ts"],
 	"references": [{ "path": "./tsconfig.node.json" }]
 }
